refactor(FON-468): extract observations feature from shared
