+++ apps/client/src/components/shared/nomination-files-table/NominationFilesTable.tsx
@@ -5,10 +5,10 @@ import './NominationFilesTable.css';
 
 import { defineMessage } from 'react-intl';
 
-import { ObservationFollowUpCommentProvider } from '../observations/follow-up-selector/ObservationFollowUpCommentDialogProvider';
 import { AlertsProvider } from '@/components/shared/alerts/AlertsProvider';
 import { DataTable, useDataTable, useQueryDataTableState } from '@/components/shared/data-table';
 import { useIsSgNavigation } from '@/features/auth/hooks/roles.hook';
+import { ObservationFollowUpCommentProvider } from '@/features/observations/context/ObservationFollowUpCommentDialogProvider';
 import type { FormationEnum, NominationFileOutcomeEnum, PrioriteEnum } from '@/types/enums.types';
 import {
   useSessionNominationFilesQuery,
