+++ apps/client/src/components/shared/nomination-files-table/components/cells/observations/ObservantsCell.tsx
@@ -1,8 +1,8 @@
 import Button from '@codegouvfr/react-dsfr/Button';
 
 import { useNominationFilesTable } from '@/components/shared/nomination-files-table/contexts/files-table.context';
-import { ObservationLinks } from '@/components/shared/ObservationLinks';
 import { useIsSgNavigation } from '@/features/auth/hooks/roles.hook';
+import { ObservationLinks } from '@/features/observations/components/ObservationLinks';
 import type { SessionNominationFile } from '@queries/nomination-sessions.queries';
 
 import { useObservationsModal } from './ObservationsModalContext';
