--- apps/client/src/components/shared/observations/index.ts
@@ -1,1 +1,0 @@
-export { ObservationCard } from './ObservationCard';
