+++ apps/client/src/features/observations/components/ObservationCard.tsx
